backend/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
scikit-learn
pandas
joblib
requests
kafka-python
numpy
sqlalchemy
pymysql
redis
gunicorn
PyJWT
bcrypt
flask-jwt-extended
